fixeddd

+++ contract/contracts/CoitonERC20.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
+
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
@@ -8,7 +9,7 @@ contract CoitonERC20 is ERC20, Ownable {
         _mint(owner, 100000000 * 10 ** decimals());
     }
 
-    function mintTo(address to, uint amount) external onlyOwner {
+    function mintTo(address to, uint256 amount) external onlyOwner {
         _mint(to, amount);
     }
 }
